+++ buffer-codec-processor/src/main/kotlin/com/ditchoom/buffer/codec/processor/CodecEmitter.kt
@@ -746,6 +746,7 @@ internal class CodecEmitter(
             is FieldSpec.ValueClassScalar -> appendDecodeValueClassScalar(body, field)
             is FieldSpec.Conditional -> appendDecodeConditional(body, field)
             is FieldSpec.ProtocolMessageScalar -> appendDecodeProtocolMessageScalar(body, field)
+            is FieldSpec.EnumScalar -> appendDecodeEnum(body, field)
         }
     }
 
@@ -828,6 +829,7 @@ internal class CodecEmitter(
             is FieldSpec.ValueClassScalar -> appendEncodeValueClassScalar(body, field)
             is FieldSpec.Conditional -> appendEncodeConditional(body, field)
             is FieldSpec.ProtocolMessageScalar -> appendEncodeProtocolMessageScalar(body, field)
+            is FieldSpec.EnumScalar -> appendEncodeEnum(body, field)
         }
     }
 
@@ -1788,6 +1790,7 @@ internal class CodecEmitter(
             is FieldSpec.ValueClassScalar -> field.valueClassType
             is FieldSpec.Conditional -> field.nullableTypeName
             is FieldSpec.ProtocolMessageScalar -> field.fieldType
+            is FieldSpec.EnumScalar -> field.enumType
         }
 }